Exclude FreeCol test suites from P/R summaries
